requirements.txt
torch
transformers
sentencepiece
python-dotenv
huggingface-hub
python-telegram-bot
nest_asyncio